third_party/WebKit/LayoutTests/inspector/css_tracker/highlight-in-source.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="../../http/tests/inspector/sources-test.js"></script>

<link rel="stylesheet" type="text/css" href="resources/highlight-in-source.css">
<script>

var initialize_CSSTracker = function() {
    InspectorTest.preloadModule("css_tracker");
}

function test()
{
    InspectorTest.runTestSuite([
        function testMarking(next)
        {            
            var tracker = self.runtime.sharedInstance(CSSTracker.CSSTrackerView);
            InspectorTest.addSniffer(CSSTracker.CSSTrackerView.prototype, "_renderRuleUsage", printResults);
    
            tracker._toggleRecording(true);
            UI.viewManager.showView("css_tracker");
            tracker._toggleRecording(false);

            function printResults()
            {
                UI.inspectorView.showPanel("sources").then(showSource);
            }
        
            function showSource()
            {
                InspectorTest.showScriptSource("highlight-in-source.css", waitForDecorations);
            }

            function waitForDecorations()
            {
                InspectorTest.addSniffer(CSSTracker.CSSTrackerView.LineDecorator.prototype, "decorate", didShowDecorations);
            }

            function didShowDecorations(sourceFrame)
            {
                var lines = Array.prototype.map.call(document.querySelectorAll(".text-editor-css-rule-unused-marker"), 
                                                        e => e.parentElement.previousSibling.textContent);

                InspectorTest.addResult(lines);   
                next();
            }
        }
    ]);
}
</script>
</head>
<p id="id">PASS</p>
<body onload="runTest()">
<p>Tests the CSS highlight in sources after the timeline recording finishes.</p>
</body>
</html>
