third_party/WebKit/LayoutTests/inspector/css_tracker/highlight-in-source-expected.txt

PASS

Tests the CSS highlight in sources after the timeline recording finishes.


Running: testMarking
5,6,7,9,10,11,12,18,19,20,22,23,24
